services/face-recognition/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
slowapi==0.1.9
pydantic>=2.5.0
structlog==23.1.0
pytest==7.4.0
pytest-asyncio==0.21.1
deepface==0.0.79
numpy>=1.22
Pillow==10.0.1
tensorflow>=2.15.0
typing-extensions>=4.8.0
prometheus-client==0.19.0
httpx==0.25.2
pymilvus==2.3.4